section 23 loader still not working

+++ admin/includes/view_all_comments.php
@@ -71,7 +71,7 @@
 
             echo "<td><a href='comments.php?approve=$comment_id'>Approve</a></td>";
             echo "<td><a href='comments.php?reject=$comment_id'>Reject</a></td>";
-            echo "<td><a href='comments.php?delete=$comment_id'>Delete</a></td>";
+            echo "<td><a onClick=\"javascript:return confirm('Are you sure you want to delete?')\" href='comments.php?delete=$comment_id'>Delete</a></td>";
             echo "</tr>";
         }
         ?>
